kernel: Add a standalone task to extract kernel image and modules
RALLY: US44069

The task extract kernel image and kernel modules to a dir specified by user
via passing KERNEL_EXTRACTDIR to local.conf or the default dir.

It should be run after compile and compile_kernelmodules, and it is not a
default task in kernel build. It can be used to extact kernel build output
quickly without going through full packaging and deployment.

Signed-off-by: Yang Shi <yang.shi@windriver.com>