kernel-dev: handle the case linux-windriver may not be the preferred linux provider
Issue: LIN7-640

There are cases that linux-windriver is not always the preferred kernel
provider even with wr-kernel layer involved, for instance, the end user
may choose a dummy kernel in their own as the kernel in use, and in which
case, the linux-windriver is not proper to be the linux-libc-headers'
provider neither.

We distinguish this situation in kernel-dev with a conditional call.

Signed-off-by: Ming Liu <ming.liu@windriver.com>