kernel: drop autogenerated commits creating obsolete data
Issue: LIN7-2764

These autogenerated files break the ability for the do_patch()
rule to quickly determine that there is no work to do, since
they dirty the end of the branch with obsolete data that
carries back to pre-yocto based builds.

Drop them so that patchless BSPs will remain in common with
their sourced counterparts, easier viewing of the tree with
git describe works as expected, and do_patch can be run
over again without a huge performance penalty.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>