hwdata: use its own database
Issue: LIN8-948

A previous patch more than three years ago removes the database of hwdata
and adds symlinks regardless the destination exists or not. It does not state
why, but chances are that it's doing this to try to use the new database.

The related codes now result in dead symlinks on target.

Now we have upgraded hwdata to the newest version, its database is new.
So there's no reason why we do not use its own database.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>